requirements.txt
absl-py==1.3.0
addict==2.4.0
alabaster==0.7.12
appdirs==1.4.4
asttokens==2.2.1
Babel==2.11.0
backcall==0.2.0
blessed==1.19.1
cachetools==5.2.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
docker-pycreds==0.4.0
docutils==0.19
executing==1.2.0
ExifRead==3.0.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.24.0
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.1
mat73==0.60
matplotlib==3.7.0
matplotlib-inline==0.1.6
networkx==3.0
numpy==1.24.1
# Jittor will auto-install CUDA dependencies
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
recommonmark==0.7.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==6.1.2
sphinx-intl==2.0.1
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==0.5.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib.applehelp==1.0.3
stack-data==0.6.2
tabulate==0.9.0
tb-nightly==2.12.0a20230109
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.10.10
tomli==2.0.1
# Replace PyTorch with Jittor
jittor>=1.3.0
# torchvision functionality will be replaced with custom implementations
einops==0.7.0
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.5
wcwidth==0.2.5
Werkzeug==2.2.2
yapf==0.32.0
zipp==3.11.0
